Changes for version 1.2
The major change is the addition of doInitialFull() to scan the full
2^32 possibilities. With some optimized routines and enhanced threading,
this makes it possible (although still slow) to operate without a starting
seed. Added a new config param to allow customization of this behavior,
although the default should work well.

Also, several minor bug fixes and cleanups that ought to have been in
their own commits.